==> scripts/4_cluster.ipynb <==
import os
import csv
import time
import pickle
import numpy as np
import pandas as pd
from tqdm import tqdm
from sentence_transformers import SentenceTransformer, util
# import networkx as nx
# from pyvis.network import Network

# model = SentenceTransformer('all-MiniLM-L6-v2')
model = SentenceTransformer('all-mpnet-base-v2')

# Read Data and Encode

all_posts_df = pd.read_csv('../data/merged_q/all_posts_max_len_40.csv')
all_replies_df = pd.read_csv('../data/merged_q/all_replies_max_len_40.csv')
print('# Posts: {}, # Replies: {}'.format(all_posts_df.shape[0], all_replies_df.shape[0]))

all_posts_text = all_posts_df['text'].tolist()
all_replies_text = all_replies_df['text'].tolist()

all_posts_embed = model.encode(all_posts_text, batch_size = 128, show_progress_bar = True, convert_to_numpy = True)

all_posts_embed.shape

all_replies_embed = model.encode(all_replies_text, batch_size = 128, show_progress_bar = True, convert_to_numpy = True)

all_replies_embed.shape

np.save('../data/merged_q/all_posts_max_len_40_embed.npy', all_posts_embed)
np.save('../data/merged_q/all_replies_max_len_40_embed.npy', all_replies_embed)

# Clustering

def community_detection(embeddings, threshold=0.75, min_community_size=10, init_max_size=1000):
    """
    Function for Fast Community Detection
    Finds in the embeddings all communities, i.e. embeddings that are close (closer than threshold).
    Returns only communities that are larger than min_community_size. The communities are returned
    in decreasing order. The first element in each list is the central point in the community.
    """

    # Compute cosine similarity scores
    cos_scores = util.pytorch_cos_sim(embeddings, embeddings)

    # Minimum size for a community
    top_k_values, _ = cos_scores.topk(k=min_community_size, largest=True)

    # Filter for rows >= min_threshold
    extracted_communities = []
    for i in range(len(top_k_values)):
        if top_k_values[i][-1] >= threshold:
            new_cluster = []

            # Only check top k most similar entries
            top_val_large, top_idx_large = cos_scores[i].topk(k=init_max_size, largest=True)
            top_idx_large = top_idx_large.tolist()
            top_val_large = top_val_large.tolist()

            if top_val_large[-1] < threshold:
                for idx, val in zip(top_idx_large, top_val_large):
                    if val < threshold:
                        break

                    new_cluster.append(idx)
            else:
                # Iterate over all entries (slow)
                for idx, val in enumerate(cos_scores[i].tolist()):
                    if val >= threshold:
                        new_cluster.append(idx)

            extracted_communities.append(new_cluster)

    # Largest cluster first
    extracted_communities = sorted(extracted_communities, key=lambda x: len(x), reverse=True)

    # Step 2) Remove overlapping communities
    unique_communities = []
    extracted_ids = set()

    for community in extracted_communities:
        add_cluster = True
        for idx in community:
            if idx in extracted_ids:
                add_cluster = False
                break

        if add_cluster:
            unique_communities.append(community)
            for idx in community:
                extracted_ids.add(idx)

    return unique_communities

# all_events_clusters = community_detection(all_events_embed, min_community_size=2, threshold=0.8)
all_posts_clusters = util.community_detection(all_posts_embed, min_community_size=2, threshold=0.8)

# with open('all_events_clusters.pickle', 'rb') as f:
#     all_events_clusters = pickle.load(f)
with open('../data/merged_q/all_posts_max_len_10_clusters.pickle', 'wb') as f:
    pickle.dump(all_posts_clusters, f)

len(all_posts_clusters)

# for i, cluster in enumerate(all_events_clusters):
#     print("\nCluster {}, #{} Elements ".format(i+1, len(cluster)))
#     for j in cluster:
#         print("\t", all_events_df.iloc[j]['text'])

# all_replies_clusters = community_detection(all_replies_embed, min_community_size=2, threshold=0.8)

with open('all_replies_clusters.pickle', 'rb') as f:
    all_replies_clusters = pickle.load(f)

len(all_replies_clusters)

# for i, cluster in enumerate(all_replies_clusters):
#     print("\nCluster {}, #{} Elements ".format(i+1, len(cluster)))
#     for j in cluster:
#         print("\t", all_replies_df.iloc[j]['text'])

num_posts_clustered = 0
for i in all_posts_clusters:
    num_posts_clustered += len(i)
print('Num posts clustered = {}'.format(num_posts_clustered))
print('Remaining num posts = {}'.format(all_posts_df.shape[0] - num_posts_clustered))
print('Total num post nodes = {}'.format(all_posts_df.shape[0] - num_posts_clustered + len(all_posts_clusters)))

cluster_post_ids = []
for i in all_posts_clusters:
    tmp_list = []
    for j in i:
        tmp_list.append(all_posts_df.iloc[j]['id'])
    cluster_post_ids.append(tmp_list)

all_posts_clusters_dict = {'cluster_id': ['post_cluster_' + str(i) for i in range(len(all_posts_clusters))],
                           'text': [all_posts_df.iloc[i[0]]['text'] for i in all_posts_clusters],
                           'post_id': cluster_post_ids,
                           'num_posts': [len(i) for i in all_posts_clusters]}

all_posts_clusters_df = pd.DataFrame(all_posts_clusters_dict)

all_posts_clusters_df

all_posts_clusters_df.to_csv('../data/merged_q/all_posts_max_len_10_clusters.csv', index = False)

cluster_reply_ids = []
for i in all_replies_clusters:
    tmp_list = []
    for j in i:
        tmp_list.append(all_replies_df.iloc[j]['id'])
    cluster_reply_ids.append(tmp_list)
    
all_replies_clusters_dict = {'cluster_id': ['reply_cluster_' + str(i) for i in range(len(all_replies_clusters))],
                           'text': [all_replies_df.iloc[i[0]]['text'] for i in all_replies_clusters],
                           'reply_id': cluster_reply_ids,
                           'num_replies': [len(i) for i in all_replies_clusters]}

all_replies_clusters_df = pd.DataFrame(all_replies_clusters_dict)

all_replies_clusters_df

# all_replies_clusters_df.to_csv('all_replies_clusters.csv', index = False)

# Visualization

pair_dict = {}
for i in range(len(all_replies_df)):
    event_id = all_replies_df.iloc[i]['parent_id']
    reply_id = all_replies_df.iloc[i]['id']
    if event_id in pair_dict:
        pair_dict[event_id].append(reply_id)
    else:
        pair_dict[event_id] = [reply_id]

event_to_cluster_dict = {}
for i in range(len(all_events_clusters_df)):
    event_cluster_id = all_events_clusters_df.iloc[i]['cluster_id']
    event_cluster_event_ids = all_events_clusters_df.iloc[i]['event_id']
    for j in event_cluster_event_ids:
        event_to_cluster_dict[j] = event_cluster_id

len(event_to_cluster_dict)

reply_to_cluster_dict = {}
for i in range(len(all_replies_clusters_df)):
    reply_cluster_id = all_replies_clusters_df.iloc[i]['cluster_id']
    reply_cluster_reply_ids = all_replies_clusters_df.iloc[i]['reply_id']
    for j in reply_cluster_reply_ids:
        reply_to_cluster_dict[j] = reply_cluster_id

len(reply_to_cluster_dict)

cluster_connection_mat = np.zeros((len(all_events_clusters_df), len(all_replies_clusters_df)))

for i in tqdm(range(len(all_replies_df))):
    event_id = all_replies_df.iloc[i]['parent_id']
    reply_id = all_replies_df.iloc[i]['id']
    if event_id in event_to_cluster_dict and reply_id in reply_to_cluster_dict:
        event_cluster_id = int(event_to_cluster_dict[event_id].split('_')[2])
        reply_cluster_id = int(reply_to_cluster_dict[reply_id].split('_')[2])
        cluster_connection_mat[event_cluster_id, reply_cluster_id] = 1

cluster_connection_mat.shape

connect_pairs = []
for i in range(cluster_connection_mat.shape[0])[70:100]:
    for j in range(cluster_connection_mat.shape[1])[50:800]:
        if cluster_connection_mat[i, j] == 1:
            connect_pairs.append(('event_cluster_' + str(i), 'reply_cluster_' + str(j)))

len(connect_pairs)

source_text_list = []
target_text_list = []
for i in connect_pairs:
    source_text_list.append(list(all_events_clusters_df[all_events_clusters_df['cluster_id'] == i[0]]['text'])[0])
    target_text_list.append(list(all_replies_clusters_df[all_replies_clusters_df['cluster_id'] == i[1]]['text'])[0])

network_df = pd.DataFrame({
    'source': source_text_list,
    'target': target_text_list
})

network_df

G = nx.from_pandas_edgelist(network_df, source = 'source', target = 'target')

source_nodes = set(list(network_df['source']))

net = Network('1000px', '1000px')

net.from_nx(G)

node_list = []
for i in net.nodes:
    if i['id'] in source_nodes:
        i['color'] = '#f4cccd'
    else:
        i['color'] = '#cde0f1'
    node_list.append(i)
net.nodes = node_list

edge_list = []
for i in net.edges:
    i['color'] = '#162347'
    edge_list.append(i)
net.edges = edge_list

net.show('example3.html')

# source_list = []
# source_text_list = []
# target_list = []
# target_text_list = []
# for i in tqdm(range(len(all_events_clusters_df))):
#     event_flag = 0
#     event_cluster_id = all_events_clusters_df.iloc[i]['cluster_id']
#     event_cluster_text = all_events_clusters_df.iloc[i]['text']
#     event_cluster_event_ids = all_events_clusters_df.iloc[i]['event_id']
#     for eid in event_cluster_event_ids:
#         for j in range(len(all_replies_clusters_df)):
#             reply_flag = 0
#             reply_cluster_id = all_replies_clusters_df.iloc[j]['cluster_id']
#             reply_cluster_text = all_replies_clusters_df.iloc[j]['text']
#             reply_cluster_reply_ids = all_replies_clusters_df.iloc[j]['reply_id']
#             for rid in reply_cluster_reply_ids:
#                 if rid in pair_dict[eid]:
#                     source_list.append(event_cluster_id)
#                     source_text_list.append(event_cluster_text)
#                     target_list.append(reply_cluster_id)
#                     target_text_list.append(reply_cluster_text)
#                     event_flag = 1
#                     reply_flag = 1
#                     break
#             if reply_flag == 1:
#                 break
#         if event_flag == 1:
#             break
