==> scripts/8_automatic_evaluation.ipynb <==
import numpy as np
import pandas as pd
from tqdm import tqdm
from nltk import word_tokenize
from nltk.translate.bleu_score import sentence_bleu
from nltk.translate.meteor_score import meteor_score
from rouge import Rouge

# Distinct Functions

def distinct_n(sentences, n):
    all_grams = []
    for sentence in sentences:
        tokens = word_tokenize(sentence.lower())
        if n > len(tokens):
            continue
        else:
            grams = [tokens[i:(i+n)] for i in range(len(tokens) - n + 1)]
            grams = [' '.join(gram) for gram in grams]
            all_grams += grams
    return len(set(all_grams)) / len(all_grams)

# Calculate Scores

df = pd.read_csv('../data/test/result/all/all_result.csv')
df

score_dict = {}
score_cols = ['model', 'dist_1', 'dist_2', 'dist_3', 'dist_4',
              'bleu_1', 'bleu_2', 'bleu_3', 'bleu_4', 'meteor', 'rouge']
for col in score_cols:
    score_dict[col] = []

models = ['moel', 'mime', 'cem', 'blender', 'meed',
          'highest_degree', 'random', 'intent', 'follow']
score_dict['model'] = models

## Distinct Scores

for n in range(1, 5):
    scores = []
    print('Distinct-{}'.format(n))
    for model in tqdm(models):
        score = distinct_n(df[model].tolist(), n)
        scores.append(score)
    score_dict['dist_{}'.format(n)] = scores

## BLEU Scores

all_refs = []
for i in tqdm(range(df.shape[0])):
    refs = [word_tokenize(x.lower()) for x in df.iloc[i]['groundtruth'].split('\n')]
    all_refs.append(refs)

weights = [(1, 0, 0, 0), (1/2, 1/2, 0, 0), (1/3, 1/3, 1/3, 0), (1/4, 1/4, 1/4, 1/4)]

for n in range(1, 5):
    print('BLEU-{}'.format(n))
    model_scores = []
    for model in tqdm(models):
        scores = []
        for i in range(df.shape[0]):
            hyp = word_tokenize(df.iloc[i][model].lower())
            score = sentence_bleu(all_refs[i], hyp, weights = weights[n - 1])
            scores.append(score)
        model_scores.append(np.mean(scores))
    score_dict['bleu_{}'.format(n)] = model_scores

score_dict['bleu_1']

score_dict['bleu_2']

score_dict['bleu_3']

score_dict['bleu_4']

## METEOR Score

model_scores = []
for model in tqdm(models):
    scores = []
    for i in range(df.shape[0]):
        hyp = word_tokenize(df.iloc[i][model].lower())
        score = meteor_score(all_refs[i], hyp)
        scores.append(score)
    model_scores.append(np.mean(scores))
score_dict['meteor'] = model_scores

score_dict['meteor']

## ROUGE Score

rouge = Rouge()

def rouge_score(refs, hyp):
    scores_1 = []
    scores_2 = []
    scores_l = []
    for ref in refs:
        score = rouge.get_scores(hyp, ref)
        scores_1.append(score[0]['rouge-1']['f'])
        scores_2.append(score[0]['rouge-2']['f'])
        scores_l.append(score[0]['rouge-l']['f'])
    return np.mean(scores_1), np.mean(scores_2), np.mean(scores_l)

model_rouge_1_scores = []
model_rouge_2_scores = []
model_rouge_l_scores = []
for model in tqdm(models):
    rouge_1_scores = []
    rouge_2_scores = []
    rouge_l_scores = []
    for i in range(df.shape[0]):
        hyp = df.iloc[i][model].lower()
        refs = [x.lower() for x in df.iloc[i]['groundtruth'].split('\n')]
        score_1, score_2, score_l = rouge_score(refs, hyp)
        rouge_1_scores.append(score_1)
        rouge_2_scores.append(score_2)
        rouge_l_scores.append(score_l)
    model_rouge_1_scores.append(np.mean(rouge_1_scores))
    model_rouge_2_scores.append(np.mean(rouge_2_scores))
    model_rouge_l_scores.append(np.mean(rouge_l_scores))
score_dict['rouge_1'] = model_rouge_1_scores
score_dict['rouge_2'] = model_rouge_2_scores
score_dict['rouge_l'] = model_rouge_l_scores

# Generate CSV

score_df = pd.DataFrame(score_dict)
score_df

score_df.to_csv('../data/test/result/all/auto_eval_result.csv', index = False)
