==> scripts/9_calculate_edges.ipynb <==
import os
import random
import numpy as np
import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt
from collections import Counter
import pickle

all_posts_df = pd.read_csv('../data/merged_q/all_posts_max_len_40.csv')
all_replies_df = pd.read_csv('../data/merged_q/all_replies_max_len_40.csv')

print('Avg speaker len:', np.mean(all_posts_df['length'].values))

print('Avg listener len:', np.mean(all_replies_df['length'].values))

all_posts_df.head()

all_replies_df.head()

print(all_posts_df.shape, all_replies_df.shape)

with open('../data/merged_q/all_posts_max_len_40_clusters/all_posts_clusters.pickle', 'rb') as f:
    all_posts_clusters = pickle.load(f)
print(len(all_posts_clusters[0.85]))

with open('../data/merged_q/all_replies_max_len_40_clusters/all_replies_clusters_combine_1_centroid.pickle', 'rb') as f:
    all_replies_clusters = pickle.load(f)
print(len(all_replies_clusters[0.8]))

num = 0
a = []
for c in all_replies_clusters[0.8]:
    num += len(c)
    a += c
print(num, len(set(a)))
print(all_replies_df.shape[0] - num + len(all_replies_clusters[0.8]))

num = 0
a = []
for c in all_posts_clusters[0.85]:
    num += len(c)
    a += c
print(num, len(set(a)))
print(all_posts_df.shape[0] - num + len(all_posts_clusters[0.85]))

print('Number of post nodes: 134061')
print('Number of reply nodes: 666587')

post_idx_to_cluster_id = {}
for i, c in enumerate(all_posts_clusters[0.85]):
    for idx in c:
        post_idx_to_cluster_id[idx] = i
print(len(post_idx_to_cluster_id))

cluster_id = len(all_posts_clusters[0.85])
for idx in range(all_posts_df.shape[0]):
    if idx not in post_idx_to_cluster_id:
        post_idx_to_cluster_id[idx] = cluster_id
        cluster_id += 1
print(len(post_idx_to_cluster_id), cluster_id)

reply_idx_to_cluster_id = {}
for i, c in enumerate(all_replies_clusters[0.8]):
    for idx in c:
        reply_idx_to_cluster_id[idx] = i
print(len(reply_idx_to_cluster_id))

cluster_id = len(all_replies_clusters[0.8])
for idx in range(all_replies_df.shape[0]):
    if idx not in reply_idx_to_cluster_id:
        reply_idx_to_cluster_id[idx] = cluster_id
        cluster_id += 1
print(len(reply_idx_to_cluster_id), cluster_id)

post_id_to_post_idx = {}
for idx in range(all_posts_df.shape[0]):
    post_id = all_posts_df.iloc[idx]['id']
    post_id_to_post_idx[post_id] = idx

adj = np.zeros((134061, 666587), dtype=bool)

for reply_idx in tqdm(range(all_replies_df.shape[0])):
    post_id = all_replies_df.iloc[reply_idx]['parent_id']
    post_idx = post_id_to_post_idx[post_id]
    adj[post_idx_to_cluster_id[post_idx], reply_idx_to_cluster_id[reply_idx]] = 1

print(np.sum(adj))

print('Avg degree of speaker nodes:', np.mean(np.sum(adj, axis = 1)))

print('Avg degree of listener nodes:', np.mean(np.sum(adj, axis = 0)))

# Profanity Check

import os
import random
import numpy as np
import pandas as pd
from tqdm import tqdm
import pickle
from profanity_check import predict, predict_prob

all_posts_df = pd.read_csv('../data/merged_q/all_posts_max_len_40.csv')
all_replies_df = pd.read_csv('../data/merged_q/all_replies_max_len_40.csv')

profanity_idx_speaker = []
for i in tqdm(range(all_posts_df.shape[0])):
    profanity = predict([all_posts_df.iloc[i]['text']])
    if profanity == 1:
        profanity_idx_speaker.append(i)
print(len(profanity_idx_speaker), all_posts_df.shape[0])

np.save('../data/merged_q/profanity_check/profanity_idx_speaker.npy', np.array(profanity_idx_speaker))

all_posts_df.iloc[profanity_idx_speaker]

profanity_idx_listener = []
for i in tqdm(range(all_replies_df.shape[0])):
    profanity = predict([all_replies_df.iloc[i]['text']])
    if profanity == 1:
        profanity_idx_listener.append(i)
print(len(profanity_idx_listener), all_replies_df.shape[0])

np.save('../data/merged_q/profanity_check/profanity_idx_listener.npy', np.array(profanity_idx_listener))

22024/838785

2526/152680

## Calculate New Nodes and Edges

import os
import random
import numpy as np
import pandas as pd
from tqdm import tqdm
import pickle

all_posts_df = pd.read_csv('../data/merged_q/all_posts_max_len_40.csv')
all_replies_df = pd.read_csv('../data/merged_q/all_replies_max_len_40.csv')

profanity_idx_speaker = np.load('../data/merged_q/profanity_check/profanity_idx_speaker.npy')
profanity_idx_listener = np.load('../data/merged_q/profanity_check/profanity_idx_listener.npy')

# cache = set(profanity_idx_speaker)
# post_idx_filtered = [i for i in range(all_posts_df.shape[0]) if i not in cache]
# print(len(post_idx_filtered))

cache = set(profanity_idx_listener)
profane_speaker_id = set([all_posts_df.iloc[i]['id'] for i in profanity_idx_speaker])
reply_idx_filtered = []
for i in tqdm(range(all_replies_df.shape[0])):
    if i in cache or all_replies_df.iloc[i]['parent_id'] in profane_speaker_id:
        continue
    reply_idx_filtered.append(i)
print(len(reply_idx_filtered))

cache = set(profanity_idx_speaker)
good_speaker_id = set(all_replies_df.iloc[reply_idx_filtered]['parent_id'].tolist())
post_idx_filtered = []
for i in tqdm(range(all_posts_df.shape[0])):
    if i in cache:
        continue
    if all_posts_df.iloc[i]['id'] not in good_speaker_id:
        continue
    post_idx_filtered.append(i)
print(len(post_idx_filtered))

# np.save('../data/merged_q/profanity_check/post_idx_filtered.npy', np.array(post_idx_filtered))
# np.save('../data/merged_q/profanity_check/reply_idx_filtered.npy', np.array(reply_idx_filtered))

# post_df_filtered = all_posts_df.iloc[post_idx_filtered]
# post_df_filtered

# reply_df_filtered = all_replies_df.iloc[reply_idx_filtered]
# reply_df_filtered

with open('../data/merged_q/all_posts_max_len_40_clusters/all_posts_clusters.pickle', 'rb') as f:
    all_posts_clusters = pickle.load(f)
print(len(all_posts_clusters[0.85]))

with open('../data/merged_q/all_replies_max_len_40_clusters/all_replies_clusters_combine_1_centroid.pickle', 'rb') as f:
    all_replies_clusters = pickle.load(f)
print(len(all_replies_clusters[0.8]))

post_cache = set(post_idx_filtered)
post_clusters = []
for c in tqdm(all_posts_clusters[0.85]):
    cc = [idx for idx in c if idx in post_cache]
    if len(cc) > 0:
        post_clusters.append(cc)
print(len(post_clusters))

reply_cache = set(reply_idx_filtered)
reply_clusters = []
for c in tqdm(all_replies_clusters[0.8]):
    cc = [idx for idx in c if idx in reply_cache]
    if len(cc) > 0:
        reply_clusters.append(cc)
print(len(reply_clusters))

post_idx_to_cluster_id = {}
for i, c in enumerate(post_clusters):
    for idx in c:
        post_idx_to_cluster_id[idx] = i

cluster_id = len(post_clusters)
for idx in range(all_posts_df.shape[0]):
    if idx not in post_cache:
        continue
    if idx not in post_idx_to_cluster_id:
        post_idx_to_cluster_id[idx] = cluster_id
        cluster_id += 1
print(cluster_id, '<-', len(post_idx_to_cluster_id), len(post_cache))

print(149332 - sum(len(c) for c in post_clusters) + len(post_clusters))

reply_idx_to_cluster_id = {}
for i, c in enumerate(reply_clusters):
    for idx in c:
        reply_idx_to_cluster_id[idx] = i

cluster_id = len(reply_clusters)
for idx in range(all_replies_df.shape[0]):
    if idx not in reply_cache:
        continue
    if idx not in reply_idx_to_cluster_id:
        reply_idx_to_cluster_id[idx] = cluster_id
        cluster_id += 1
print(cluster_id, '<-', len(reply_idx_to_cluster_id), len(reply_cache))

print(803320 - sum(len(c) for c in reply_clusters) + len(reply_clusters))

post_id_to_post_idx = {}
for idx in tqdm(range(all_posts_df.shape[0])):
    if idx not in post_cache:
        continue
    post_id = all_posts_df.iloc[idx]['id']
    post_id_to_post_idx[post_id] = idx

adj = np.zeros((131038, 637628), dtype=bool)

for reply_idx in tqdm(range(all_replies_df.shape[0])):
    if reply_idx not in reply_cache:
        continue
    post_id = all_replies_df.iloc[reply_idx]['parent_id']
    post_idx = post_id_to_post_idx[post_id]
    adj[post_idx_to_cluster_id[post_idx], reply_idx_to_cluster_id[reply_idx]] = 1

print('Number of edges:', np.sum(adj))

print('Avg degree of speaker nodes:', np.mean(np.sum(adj, axis = 1)))

print('Avg degree of listener nodes:', np.mean(np.sum(adj, axis = 0)))

print('Number of speaker nodes: 131038')
print('Number of listener nodes: 637628')
